pyproject.toml
[project]
name = "ai-engineer-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.95.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
]
